+++ src/realtime/index.ts
@@ -1,2 +1,2 @@
 export * from "./client/index";
-export * from './server/index';
+export * from "./server/index";