+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 Most patterns are highly stable, no changes will be made to existing methods, only extended, but I will adhere to adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) just in case. You can safely upgrade, but as always, RTFM (read changelog for major releases).
 
+
+## [2.1.0] - 2025-06-17
+
+### Added:
+
+- Realtime Server
+- Realtime Client (streamlined version of RealtimeService)
+- APIResponse, APIError patterns
+- Many Realtime examples: Realtime Client (NATS, Websocket), Realtime Server (NATS, Websocket)
+
 ## [2.0.3] - 2025-06-14
 
 ### Added
@@ -38,8 +48,7 @@ Not for production.
 
 - Improved tests
 - Refactored
-- Streamlined based on feedback (nothing broken) 
-
+- Streamlined based on feedback (nothing broken)
 
 ## [1.1.7] - 2025-05-21